ch04/requirements.txt
transformers==4.35.0
torch==2.1.0
Pillow==10.0.1
requests==2.31.0
huggingface-hub==0.19.0
Flask==2.3.3
